+++ src/api/formatters.ts
@@ -10,12 +10,16 @@ import path from 'path'
 import { DEFAULT_CUCUMBER_PUBLISH_URL } from '../formatter/publish'
 import HttpStream from '../formatter/http_stream'
 import { Writable } from 'stream'
+import { supportsColor } from 'supports-color'
 import { IRunOptionsFormats } from './types'
+import hasAnsi from 'has-ansi'
+import stripAnsi from 'strip-ansi'
 
 export async function initializeFormatters({
   env,
   cwd,
   stdout,
+  stderr,
   logger,
   onStreamError,
   eventBroadcaster,
@@ -26,6 +30,7 @@ export async function initializeFormatters({
   env: NodeJS.ProcessEnv
   cwd: string
   stdout: IFormatterStream
+  stderr: IFormatterStream
   logger: Console
   onStreamError: () => void
   eventBroadcaster: EventEmitter
@@ -88,7 +93,7 @@ export async function initializeFormatters({
     const readerStream = new Writable({
       objectMode: true,
       write: function (responseBody: string, encoding, writeCallback) {
-        logger.error(responseBody)
+        logger.error(sanitisePublishOutput(responseBody, stderr))
         writeCallback()
       },
     })
@@ -99,4 +104,17 @@ export async function initializeFormatters({
   return async function () {
     await Promise.all(formatters.map(async (f) => await f.finished()))
   }
+}
+
+/*
+This is because the Cucumber Reports service returns a pre-formatted console message
+including ANSI escapes, so if our stderr stream doesn't support those we need to
+strip them back out. Ideally we should get structured data from the service and
+compose the console message on this end.
+ */
+function sanitisePublishOutput(raw: string, stderr: IFormatterStream) {
+  if (!supportsColor(stderr) && hasAnsi(raw)) {
+    return stripAnsi(raw)
+  }
+  return raw
 }